+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Unreleased
 - fields.BlockField: Fix validation of exists flag
 - fields.StreamField: Add end_offset as an alternative to length
 - fields.ChunkField: Add length_before_id option
+- refs.Coda: Add experimental feature for attaching Fields to the end of Blocks
 
 0.9.0 - 2021-01-14
 ==================